==> 05-enunciat.Rmd <==
---
title: "Treball d'Estadística Aplicada"
output: 
  pdf_document: 
    extra_dependencies: wrapfig
    toc: yes
header-includes:
  - \usepackage[catalan]{babel}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
library(readr)
library(ggplot2)
library(lmtest)
library(readxl)
if(!exists(".ID")) .ID = '000000'
penguins = read_excel( sprintf("05-enunciat/penguins_%s.xlsx", .ID))
url_dades = sprintf("https://moodle2.udg.edu/pluginfile.php/1890669/mod_folder/content/0/penguins_%s.xlsx", .ID)
```

# El conjunt de dades `penguins`

El conjunt `penguins` conté _vuit_ característiques de diferents espècies de pingüins. El nom de les columnes són:

* `species`: Espècie del pingüí: [pingüí de Papua](https://ca.wikipedia.org/wiki/Ping%C3%BC%C3%AD_de_corona_blanca) ("Gentoo"), [pingüí d'Adèlia](https://ca.wikipedia.org/wiki/Ping%C3%BC%C3%AD_de_corona_blanca) ("Adelie") i [pingüí carablanc](https://ca.wikipedia.org/wiki/Ping%C3%BC%C3%AD_carablanc) ("Chinstrap").
* `island`: illa on s'ha registrat l'exemplar: illes Biscoe, illa Dream i illa Torgersen.
* `bill_length_mm`: llargada del bec en mil·límetres.
* `bill_depth_mm`: profunditat del bec en mil·límetres.
* `flipper_length_mm`: llargada de l'aleta en mil·límetres.
* `body_mass_g`: pes en grams.
* `sex`: sexe del pingüí.
* `year`: any que s'ha registrat.

Trobareu el conjunt de dades a l'arxiu `penguins_`r .ID`.xlsx` a: 

<`r url_dades`>

Si no us funciona l'enllaç, també trobareu l'arxiu de dades al llistat de grups del Moodle.


Els apartats que haurà de contenir el treball són:

* __Descripció del conjunt de dades__: breu resum de les dades
* __Sobre el bec dels pingüins__: descripció d'una relació dels becs dels pingüins
* __Sobre el sexe dels pingüins__: descripció de la relació del sexe amb altres característiques dels pingüins.
* __Sobre els pingüins d'Adèlia__: descripció del pes dels pingüins d'Adèlia trobats en diferents zones.

A continuació es detalla què cal fer en cadascun d'aquests apartats.

## Descripció del conjunt de dades

El treball haurà de començar amb una introducció on s'expliqui el conjunt de dades amb el que s'està treballant. Aquesta introducció haurà de tenir una taula descriptiva del conjunt amb què s'està treballant. Podeu trobar molts exemples de taules descriptives escrivint el text "descriptive table" a Google.

<p style="color:blue;">__Indicacions per construir la taula__. Si la variable és numèrica, caldrà mostrar un estadístic de tendència central i un estadístic de dispersió. Si la variable és categòrica, per cada categoria caldrà mostrar la freqüència absoluta i la freqüència relativa.</p>

## Sobre el bec dels pingüins

En el nostre conjunt de dades hem observat que la llargada del bec és aproximadament 2.5 vegades més gran que la profunditat del bec. Voldríem veure si aquesta hipòtesi és plausible.

* Construïu una nova variable anomenada `bill_depth_mm_2.5` multiplicant la variable `bill_depth_mm` per 2.5.
* Construïu una segona variable anomenada `difference` que contingui la diferència entre `bill_length_mm` i `bill_depth_mm_2.5`.
* A partir de la nova variable construïda, intenteu justificar la plausibilitat de la hipòtesis que s'està plantejant. 

<p style="color:blue;">__Objectiu d'aquest apartat__: a partir de la variable `difference`, defensar mitjançant estadística descriptiva (numèrica i gràfica) la certesa de la hipòtesi plantejada.</p>


## Sobre el sexe dels pingüins

Hem observat que el sexe dels pingüins no sembla aportar gaire informació sobre l'animal ni la mostra recollida. Ens hem plantejat tres hipòtesis sobre característiques de la zona mostrejada que voldríem analitzar:

1. La proporció de pingüins mascle i femella a la zona és semblant.
1. La profunditat del bec és semblant en funció del sexe del pingüí.
1. La distribució del sexe dels pingüins és independent de l'espècie.

<p style="color:blue;">__Objectiu d'aquest apartat__: a partir de les variables implicades en cada apartat, defensar mitjançant estadística descriptiva (numèrica i gràfica) la certesa de cada una de les hipòtesis plantejades. Si voleu, en el vostre treball, podeu dividir les tres hipotesis en tres subapartats</p>

## Sobre els pingüins d'Adèlia

\begin{wrapfigure}{l}{0.25\textwidth}
\includegraphics[width=0.9\linewidth]{adelie.jpg} 
\end{wrapfigure}

En aquest apartat únicament considerarem els pingüins d'Adèlia (`species == 'Adelie'`). Per això, primerament filtrem el conjunt dades. Això ho podem fer de la següent manera:

- Aneu al menú _Dades->Taula de dades activa->Fes un subconjunt de la taula de dades activa_.
- Seleccioneu l'opció per defecte "Inclou totes les variables".
- A _Expressió de selecció_ poseu: `species == 'Adelie'`.
- A _Nom de la nova taula de dades_ poseu el nom de la nova taula, per exemple: `penguins_adelie`.

```{r, include=FALSE}
penguins_adelie = penguins %>%
  filter(species == 'Adelie')
```

Si ho heu fet bé, obtindreu una nova taula anomenada `penguins_adelie` amb els `r nrow(penguins_adelie)` pingüins de l'espècie Adelie.

Aquest tipus de pingüí ha estat observat a diferents illes (`island`). Voldríem veure si el pes dels animals és similar en les diferents illes. 

<p style="color:blue;">__Objectiu d'aquest apartat__: a partir de les variables implicades, defensar mitjançant estadística descriptiva (numèrica i gràfica) la certesa de la hipòtesi plantejada.</p>

</div>

\newpage

# Termini i instruccions d'entrega

La primera part del treball s'ha de presentar el dia __16 d'abril__ a través de Moodle. Trobareu l'entrega dins l'apartat "TREBALL".

__Si no seguiu aquestes instruccions, pot ser que obtinguis un zero a la nota del treball.__

* És un treball on s’han de mostrar l’assoliment dels continguts, procediments i mètodes de l’assignatura.
* El treball que presenteu ha de ser anònim. El vostre nom no hi pot aparèixer, únicament pot aparèixer-hi el codi UdG de les persones del grup.
* Cal raonar els resultats i les passes seguides per aconseguir-los.
* Si ho trobeu convenient, feu una captura de pantalla amb el menú i els camps que heu utilitzat.
* Incorporeu només les sortides numèriques i gràfiques d’R que creieu indispensables acompanyades sempre d’una explicació/valoració. __No es valoraran els apartats que es limitin a copiar sortides d’R.__
* Es valorarà molt positivament la claredat i la concisió en la presentació de resultats.
* Aquest treball pràctic s’ha de lliurar a través del mòdul de Recollida de treballs i pràctiques del Moodle. No s’acceptaran els treballs pràctics que es lliurin fora del termini.

__Molt important__: cada una de les parts del treball s’han de lliurar en un arxiu pdf. __No s’acceptaran treballs amb cap altre format.__
