==> 08B-solucio.Rmd <==
---
title: "Guia per corregir el segon treball d'Estadística Aplicada"
output: 
  pdf_document: 
    extra_dependencies: wrapfig
header-includes:
  - \usepackage[catalan]{babel}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, comment = " ")
library(dplyr)
library(readr)
library(ggplot2)
library(lmtest)
library(readxl)
if(!exists(".ID")) .ID = '2002963'
penguins = read_excel( sprintf("05-enunciat/penguins_%s.xlsx", .ID))
url_dades = sprintf("https://moodle2.udg.edu/pluginfile.php/1890669/mod_folder/content/0/penguins_%s.xlsx", .ID)
```

## Sobre el bec dels pingüins

* En aquest apartat és igualment correcte analitzar la semblança de la mitjana de la llargada del bec i el producte de 2.5 per la profunditat del bec utilitzant un interval de confiança o utilitzant un contrast de mitjanes per mostres aparellades.

#### Interval de confiança {-}

* Per valorar l'interval de confiança s'ha de construir la variable `difference = bill_length_mm - 2.5 * bill_depth_mm` i mirar si l'interval de confiança de la mitjana d'aquesta variable conté o no conté el zero.

```{r}
CI = confint(lm( ( bill_length_mm - 2.5 * bill_depth_mm)~ 1, data = penguins))
rownames(CI) = 'IC 95%'
colnames(CI) = NULL
knitr::kable(CI)
```

* Tingueu en compte que si s'ha calculat la variable `difference` com `2.5 * bill_depth_mm - bill_length_mm` l'interval estarà del revès.
* Cal justificar que si l'interval conté el zero, podem assumir que les mitjanes de les variables `bill_length_mm` i `2.5 * bill_depth_mm`, mentre que si l'interval no conté el zero aleshores tenim evidències de que les mitjanes de `bill_length_mm` i `2.5 * bill_depth_mm` són diferents.

#### Contrast d'hipotesis {-}

* Pel contrast d'hipòtesi tenim dues alternatives. O s'ha fet un test dues mitjanes per dades aparellades, o s'ha fet un test d'una mitjana utilitzant la variable `difference = bill_length_mm - 2.5 * bill_depth_mm`. En el primer cas, la hipòtesi nul·la seria

\[
\text{H}_0: \mu_\text{length} = \mu_\text{2.5 $\times$depth} \text{ vs } \text{H}_1: \mu_\text{length} \neq \mu_\text{2.5 $\times$depth}
\]

si s'ha optat per fer el contrast sobre la diferència, la hipòtesi nul·la podria estar escrita com

\[
\text{H}_0: \mu_\text{difference} = 0 \text{ vs } \text{H}_1: \mu_\text{difference} \neq 0
\]

El resultat que s'obté amb R fent el test de dades aparellades per dues mitjanes és

```{r}
with(penguins, t.test(bill_length_mm, 2.5 * bill_depth_mm, paired = TRUE))
```

que seria igual que si féssim el test sobre la variable `difference`

```{r}
difference = with(penguins, bill_length_mm - 2.5 * bill_depth_mm)
t.test(difference)
```

* Cal que en el treball s'hi expliqui la conclusió del contrast en relació amb el valor $p$ observat.

## Sobre el sexe dels pingüins

### La proporció de pingüins mascle i femella a la zona és semblant. {-}

En aquest apartat s'espera que es faci un contrast de bondat d'ajust. 

* Per contrastar les hipòtesis

\[
\text{H}_0: \pi_\text{female} = 0.5 \text{ i } \pi_\text{male} = 0.5\text{ vs } \text{H}_1: \pi_\text{female} \neq 0.5 \text{ o } \pi_\text{male} \neq 0.5
\]

es pot utilitzar tant el test $\chi^2$:

```{r, echo=TRUE}
chisq.test(table(penguins$sex), p = c(0.5,0.5))
```

com el contrast G:

```{r, echo=TRUE}
library(DescTools)
GTest(table(penguins$sex), p = c(0.5,0.5))
```

* També seria vàlid fer el contrast d'una proporció mirant si algun dels sexes podem assumir que és igual a 0.5.

### La profunditat del bec és semblant en funció del sexe del pingüí. {-}

En aquest apartat volem fer un contrast de comparació de mitjanes, en un disseny de mostres independents. Primer hem de mirar si podem assumir que la variància de la profunditat del bec és la mateixa pels dos sexes. Per poder assumir-ho podem fer el test F de dues variàncies:

\[
\text{H}_0: \sigma_\text{female} = \sigma_\text{male} \text{ vs } \text{H}_1: \sigma_\text{female} \neq \sigma_\text{male}
\]

```{r}
var.test(bill_depth_mm~sex, data = penguins)
```

si veiem que les variàncies poden ser assumides iguals, haurem d'assumir-ho (i marcar l'opció variàncies igual) en contrastar les mitjanes:

\[
\text{H}_0: \mu_\text{female} = \mu_\text{male} \text{ vs } \text{H}_1: \mu_\text{female} \neq \mu_\text{male}
\]

```{r}
t.test(bill_depth_mm~sex, data = penguins, var.equal = TRUE)
```

* Per veure si realment s'ha fet un contrast assumint variàncies iguals hauríem de veure el títol "Two Sample t-test" a la sortida d'R en lloc del títol "Welch Two Sample t-test" que és el test a utilitzar quan les variàncies no poden ser assumides iguals.

### La distribució del sexe dels pingüins és independent de l'espècie. {-}

En aquest apartat cal fer un contrast d'independència:

\[
\begin{array}{l}
\text{H}_0: \text{Les variables sexe i espècie són independents}\\
\text{H}_1: \text{Existeix alguna dependència entre les variables sexe i espècie}
\end{array}
\]

```{r}
with(penguins, chisq.test(table(sex, species)))
```


## Sobre els pingüins d'Adèlia

```{r, include=FALSE}
penguins_adelie = penguins %>%
  filter(species == 'Adelie')
```

Utilitzant la taula filtrada, volem veure si podem assumir que el pes dels pingüins és el mateix a les tres illes.

\[
\text{H}_0: \mu_\text{Biscoe} = \mu_\text{Dream} = \mu_\text{Torgersen} \text{ vs } \text{H}_1: \text{Algun $\mu_i$ és diferent de la resta }
\]

Farem un test Anova. Primer necessitem veure si podem assumir variàncies iguals en els residus (homoscedasticitat), ho comprovarem utilitzant el test de Levene

\[
\text{H}_0: \text{Els residus són homoscedastics  vs } \text{H}_1: \text{Els residus són heteroscedastics}
\]

```{r}
DescTools::LeveneTest(body_mass_g~factor(island), data=penguins_adelie)
```

Un cop comprovat, realitzarem el contrast Anova assumint o no variàncies iguals.

* Assumint variàncies iguals:

```{r}
oneway.test(body_mass_g~factor(island), data=penguins_adelie, var.equal = TRUE)
```

* No assumint variàncies iguals:

```{r}
oneway.test(body_mass_g~factor(island), data=penguins_adelie, var.equal = FALSE)
```

Finalment, podem contrastar la normalitat utilitzant el test de Shapiro-Wilk

\[
\text{H}_0: \text{Tenim normalitat als residus  vs } \text{H}_1: \text{No tenim normalitat als residus}
\]

```{r}
residus = residuals(m <- lm(body_mass_g~factor(island), data=penguins_adelie))
shapiro.test(residus)
```

* Donarem per vàlid una justificació dels supòsits utilitzant els gràfics de diagnòstic

```{r, fig.width=6, fig.height=5, out.width="50%", fig.align='center'}
cpar = par(mfrow=c(2,2))
plot(m)
par(cpar)
```
