==> 09B-enunciat_correccio.Rmd <==
---
title: "Dossier de correcció"
output:
  html_document:
    self_contained: yes
    mode: selfcontained
---

```{r setup, include=FALSE}
library(tidyverse)
library(knitr)
knitr::opts_chunk$set(echo = FALSE)
if(!exists('.ID')){
  .ID = first(scan('udg_codis_correctors.txt', what = 'character'))
}
source('00-config.R')
load('07B-correccio.RData')

assignment = filter(assignments, id == .ID)

id = assignment$id
treb1 = as.character(assignment$treb1)
treb2 = as.character(assignment$treb2)
treb3 = as.character(assignment$treb3)

link_form = function(id, id_treb){
  sprintf("https://docs.google.com/forms/d/e/1FAIpQLSdYzvv-lPwpSJdIxCYZeocHYpcwi9U-hedd6FElArE110HezQ/viewform?usp=pp_url&entry.1073457589=%s&entry.1570597957=%s", id, id_treb)
  # sprintf("https://docs.google.com/forms/d/e/1FAIpQLSenxYjXCLp8x6cu0lOuIiECg09f8WiA5o9FsJSPB5VDLe--Aw/viewform?usp=pp_url&entry.1073457589=%s&entry.1570597957=%s", id, id_treb)
}
treballs = pivot_longer(assignment, matches("treb."), names_to = 'name2', values_to = 'id_tr')
```


Aquest document s'ha generat pel codi d'estudiant: `r id`. Si us plau, abans de seguir endavant comprova que és el teu.

A la següent pàgina trobaràs informació de tres treballs realitzats per altres grups. De cada treball trobaràs enllaços al treball presentat, al formulari de correcció, a l'enunciat original que tenia el treball i a un document amb solucions del treball. Per accedir als enllaços hauries de poder clicar a sobre. Altrament, pots copiar i enganxar les adreces en un navegador. Per alguns dels enllaços caldrà que entris la teva informació per poder accedir al Moodle de l'assignatura.

Hauràs de revisar els treballs de grups: __`r treb1`__, __`r treb2`__ i __`r treb3`__. 

Finalment, recorda que el termini per realitzar la correcció és el dia 11 de juny.

\newpage

# Avaluació del treballs

### Treball 1 (`r treb1`)

```{r}
id_tr = as.character(slice(treballs, 1)$id_tr)
url_treball = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/treball_2_%s.pdf", ID_COURSE, id_tr)
url_formulari = link_form(id, id_tr)
url_enunciat = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/enunciat2_%s.pdf", ID_COURSE, id_tr)
url_dades = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/penguins_%s.xlsx", ID_COURSE, id_tr)
url_solucio = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/solucio2_%s.pdf", ID_COURSE, id_tr)
```

* Identitat del treball:
__`r id_tr`__

* Treball a avaluar: [pdf](`r url_treball`)


* Formulari de correcció del treball: [Google Forms](`r url_formulari`)


* Enunciat original del treball i dades del treball: [pdf](`r url_enunciat`), [xlsx](`r url_dades`)


* Solució proposada del treball: [pdf](`r url_solucio`)

### Treball 2 (`r treb2`)

```{r}
id_tr = as.character(slice(treballs, 2)$id_tr)
url_treball = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/treball_2_%s.pdf", ID_COURSE, id_tr)
url_formulari = link_form(id, id_tr)
url_enunciat = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/enunciat2_%s.pdf", ID_COURSE, id_tr)
url_dades = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/penguins_%s.xlsx", ID_COURSE, id_tr)
url_solucio = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/solucio2_%s.pdf", ID_COURSE, id_tr)
```

* Identitat del treball:
__`r id_tr`__

* Treball a avaluar: [pdf](`r url_treball`)


* Formulari de correcció del treball: [Google Forms](`r url_formulari`)


* Enunciat original del treball i dades del treball: [pdf](`r url_enunciat`), [xlsx](`r url_dades`)


* Solució proposada del treball: [pdf](`r url_solucio`)

### Treball 3 (`r treb3`)

```{r}
id_tr = as.character(slice(treballs, 3)$id_tr)
url_treball = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/treball_2_%s.pdf", ID_COURSE, id_tr)
url_formulari = link_form(id, id_tr)
url_enunciat = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/enunciat2_%s.pdf", ID_COURSE, id_tr)
url_dades = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/penguins_%s.xlsx", ID_COURSE, id_tr)
url_solucio = sprintf("https://moodle2.udg.edu/pluginfile.php/%d/mod_folder/content/0/solucio2_%s.pdf", ID_COURSE, id_tr)
```

* Identitat del treball:
__`r id_tr`__

* Treball a avaluar: [pdf](`r url_treball`)


* Formulari de correcció del treball: [Google Forms](`r url_formulari`)


* Enunciat original del treball i dades del treball: [pdf](`r url_enunciat`), [xlsx](`r url_dades`)


* Solució proposada del treball: [pdf](`r url_solucio`)
